Stop overriding ISO digits with CLDR digits.
For ~14 currencies ISO specifies 2 or 3 digits, but CLDR specifies 0,
because that better reflects real world usage, where the currency's
minor units no longer have value and aren't used in cash transactions.

However, it is common in such cases for the digits to still be used
for banking transactions, tax amounts, etc.
Since those represent common use cases for this package, it is safer to
default to the ISO digits, and let consumers further
restrict and round as needed.

Fixes #37.